# src/main/java/com/movie/catalog/controller/GenreController.java
package com.movie.catalog.controller;

import java.util.List;

import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

import com.movie.catalog.dto.GenreRequestDTO;
import com.movie.catalog.dto.GenreResponseDTO;
import com.movie.catalog.service.GenreService;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.validation.Valid;

@RestController
@RequestMapping("/api/genres")
@Tag(
    name = "Genre Management",
    description = "APIs for managing movie genres"
)
public class GenreController {

    private final GenreService genreService;

    public GenreController(GenreService genreService) {
        this.genreService = genreService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    @Operation(summary = "Create genre")
    public GenreResponseDTO createGenre(@Valid @RequestBody GenreRequestDTO dto) {
        return genreService.create(dto);
    }

    @GetMapping
    @Operation(summary = "Get all genres")
    public List<GenreResponseDTO> getAllGenres() {
        return genreService.getAll();
    }
}